+++ src/groups/group_operation_action.jl
@@ -65,12 +65,35 @@ end
 
 function adjoint_apply_diff_group(A::GroupOperationAction, a, X, p)
     G = base_group(A)
-    return inverse_translate_diff(G, a, p, X, reverse_direction_and_side(A))
+    if direction_and_side(A) === LeftForwardAction() ||
+       direction_and_side(A) === RightBackwardAction()
+        return inverse_translate_diff(G, a, p, X, reverse_direction_and_side(A))
+    else
+        return inverse_translate_diff(
+            G,
+            p,
+            a,
+            adjoint_inv_diff(G, apply(A, a, p), X),
+            (direction(A), switch_side(action_side(A))),
+        )
+    end
 end
 
 function adjoint_apply_diff_group!(A::GroupOperationAction, Y, a, X, p)
     G = base_group(A)
-    return inverse_translate_diff!(G, Y, a, p, X, reverse_direction_and_side(A))
+    if direction_and_side(A) === LeftForwardAction() ||
+       direction_and_side(A) === RightBackwardAction()
+        return inverse_translate_diff!(G, Y, a, p, X, reverse_direction_and_side(A))
+    else
+        return inverse_translate_diff!(
+            G,
+            Y,
+            p,
+            a,
+            adjoint_inv_diff(G, apply(A, a, p), X),
+            (direction(A), switch_side(action_side(A))),
+        )
+    end
 end
 
 apply(A::GroupOperationAction, a, p) = translate(A.group, a, p, direction_and_side(A))
@@ -95,13 +118,65 @@ function apply_diff!(A::GroupOperationAction, Y, a, p, X)
     return translate_diff!(A.group, Y, a, p, X, direction_and_side(A))
 end
 
+@doc raw"""
+    apply_diff_group(A::GroupOperationAction, a, X, p)
+
+Compute differential of [`GroupOperationAction`](@ref) `A` with respect to group element
+at tangent vector `X`:
+
+````math
+(\mathrm{d}τ^p) : T_{a} \mathcal G → T_{τ_a p} \mathcal G
+````
+
+There are four cases:
+* left action from the left side: ``L_a: p ↦ a \circ p``, where
+````math
+(\mathrm{d}L_a) : T_{a} \mathcal G → T_{a \circ p} \mathcal G.
+````
+* right action from the left side: ``L'_a: p ↦ a^{-1} \circ p``, where
+````math
+(\mathrm{d}L'_a) : T_{a} \mathcal G → T_{a^{-1} \circ p} \mathcal G.
+````
+* right action from the right side: ``R_a: p ↦ p \circ a``, where
+````math
+(\mathrm{d}R_a) : T_{a} \mathcal G → T_{p \circ a} \mathcal G.
+````
+* left action from the right side: ``R'_a: p ↦ p \circ a^{-1}``, where
+````math
+(\mathrm{d}R'_a) : T_{a} \mathcal G → T_{p \circ a^{-1}} \mathcal G.
+````
+
+"""
 function apply_diff_group(A::GroupOperationAction, a, X, p)
     G = base_group(A)
-    return translate_diff(G, p, a, X, reverse_direction_and_side(A))
+    if direction_and_side(A) === LeftForwardAction() ||
+       direction_and_side(A) === RightBackwardAction()
+        return translate_diff(G, p, a, X, reverse_direction_and_side(A))
+    else
+        return translate_diff(
+            G,
+            p,
+            a,
+            inv_diff(G, a, X),
+            (direction(A), switch_side(action_side(A))),
+        )
+    end
 end
 function apply_diff_group!(A::GroupOperationAction, Y, a, X, p)
     G = base_group(A)
-    return translate_diff!(G, Y, p, a, X, reverse_direction_and_side(A))
+    if direction_and_side(A) === LeftForwardAction() ||
+       direction_and_side(A) === RightBackwardAction()
+        return translate_diff!(G, Y, p, a, X, reverse_direction_and_side(A))
+    else
+        return translate_diff!(
+            G,
+            Y,
+            p,
+            a,
+            inv_diff(G, a, X),
+            (direction(A), switch_side(action_side(A))),
+        )
+    end
 end
 
 function inverse_apply_diff(A::GroupOperationAction, a, p, X)