+++ src/groups/product_group.jl
@@ -40,6 +40,18 @@ end
     end
 end
 
+function adjoint_inv_diff!(G::ProductGroup, Y, p, X)
+    M = G.manifold
+    map(
+        adjoint_inv_diff!,
+        M.manifolds,
+        submanifold_components(G, Y),
+        submanifold_components(G, p),
+        submanifold_components(G, X),
+    )
+    return Y
+end
+
 function identity_element(G::ProductGroup)
     M = G.manifold
     return ArrayPartition(map(identity_element, M.manifolds))
@@ -101,6 +113,18 @@ function inv!(G::ProductGroup, q, p)
     return q
 end
 inv!(::ProductGroup, q::Identity{ProductOperation}, ::Identity{ProductOperation}) = q
+
+function inv_diff!(G::ProductGroup, Y, p, X)
+    M = G.manifold
+    map(
+        inv_diff!,
+        M.manifolds,
+        submanifold_components(G, Y),
+        submanifold_components(G, p),
+        submanifold_components(G, X),
+    )
+    return Y
+end
 
 _compose(G::ProductGroup, p, q) = _compose(G.manifold, p, q)
 function _compose(M::ProductManifold, p::ArrayPartition, q::ArrayPartition)