+++ test/groups/circle_group.jl
@@ -46,6 +46,8 @@ using Manifolds:
             test_invariance=true,
             test_lie_bracket=true,
             test_adjoint_action=true,
+            test_inv_diff=true,
+            test_adjoint_inv_diff=true,
         )
     end
 
@@ -65,6 +67,8 @@ using Manifolds:
             test_invariance=true,
             test_lie_bracket=true,
             test_adjoint_action=true,
+            test_inv_diff=true,
+            test_adjoint_inv_diff=true,
         )
     end