+++ test/groups/special_orthogonal.jl
@@ -59,6 +59,9 @@ using Manifolds: LeftForwardAction, RightBackwardAction
                 test_exp_from_identity=true,
                 test_log_from_identity=true,
                 test_vee_hat_from_identity=true,
+                test_inv_diff=true,
+                test_adjoint_inv_diff=true,
+                test_apply_diff_group=true,
             )
 
             @testset "log_lie edge cases" begin